+++ app/assets/stylesheets/common/select-kit/_index.scss
@@ -6,7 +6,6 @@
 @import "color-palettes";
 @import "color-palette-picker";
 @import "combo-box";
-@import "composer-actions";
 @import "dropdown-select-box";
 @import "email-group-user-chooser";
 @import "flair-row";
