+++ app/assets/stylesheets/wcag.scss
@@ -44,12 +44,6 @@ html {
     color: var(--primary);
   }
 
-  // Composer
-  #reply-control .reply-to .reply-details .d-icon {
-    opacity: 1;
-    color: var(--primary-low-mid);
-  }
-
   .d-editor-button-bar {
     .btn-icon.btn-default .d-icon {
       color: var(--primary-low-mid);
