+++ config/site_settings.yml
@@ -4714,15 +4714,6 @@ experimental:
       impact: "other,all_members"
       body_class: true
       learn_more_url: "https://meta.discourse.org/t/-/407849"
-  enable_new_composer_actions:
-    default: false
-    client: true
-    hidden: true
-    area: "experimental"
-    upcoming_change:
-      status: "permanent"
-      impact: "feature,all_members"
-      learn_more_url: "https://meta.discourse.org/t/-/403635"
   prioritize_recently_used_tags:
     default: false
     client: true
