+++ db/post_migrate/20260813202233_remove_enable_new_composer_actions_setting.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+class RemoveEnableNewComposerActionsSetting < ActiveRecord::Migration[8.0]
+  def up
+    execute "DELETE FROM site_settings WHERE name = 'enable_new_composer_actions'"
+  end
+
+  def down
+    raise ActiveRecord::IrreversibleMigration
+  end
+end
