+++ frontend/discourse/tests/acceptance/composer-actions-test.js
@@ -1,13 +1,14 @@
 import { click, fillIn, visit, waitFor } from "@ember/test-helpers";
 import { test } from "qunit";
 import sinon from "sinon";
-import { cloneJSON } from "discourse/lib/object";
-import Draft from "discourse/models/draft";
 import {
   _clearRegisteredActions,
   registerComposerAction,
-} from "discourse/select-kit/components/composer-actions";
+} from "discourse/lib/composer/actions-registry";
+import { cloneJSON } from "discourse/lib/object";
+import Draft from "discourse/models/draft";
 import discoveryFixtures from "discourse/tests/fixtures/discovery-fixtures";
+import topicFixtures from "discourse/tests/fixtures/topic";
 import userFixtures from "discourse/tests/fixtures/user-fixtures";
 import {
   acceptance,
@@ -17,7 +18,26 @@ import {
 import selectKit from "discourse/tests/helpers/select-kit-helper";
 import { i18n } from "discourse-i18n";
 
-acceptance(`Composer Actions`, function (needs) {
+function composerActionsDropdown() {
+  return {
+    async expand() {
+      await click(".composer-actions-trigger");
+    },
+    async selectRowByValue(value) {
+      await click(`[data-action-id='${value}']`);
+    },
+    rows() {
+      return document.querySelectorAll(
+        ".composer-actions-dropdown [data-action-id]"
+      );
+    },
+    actionIds() {
+      return [...this.rows()].map((row) => row.dataset.actionId);
+    },
+  };
+}
+
+acceptance("Composer Actions", function (needs) {
   needs.user({
     id: 5,
     username: "kris",
@@ -39,30 +59,30 @@ acceptance(`Composer Actions`, function (needs) {
 
       return helper.response(response);
     });
+    server.get("/t/54077.json", () => {
+      const response = cloneJSON(topicFixtures["/t/54077.json"]);
+      response.fancy_title =
+        '<span dir="auto">Short topic with two posts</span>';
+
+      return helper.response(response);
+    });
   });
 
   test("replying to post", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
 
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
     await composerActions.expand();
 
-    assert.strictEqual(
-      composerActions.rowByIndex(0).value(),
-      "reply_as_new_topic"
-    );
-    assert.strictEqual(composerActions.rowByIndex(1).value(), "reply_to_topic");
-    assert.strictEqual(composerActions.rowByIndex(2).value(), "toggle_whisper");
-    assert.strictEqual(
-      composerActions.rowByIndex(3).value(),
-      "toggle_topic_bump"
-    );
-    assert.strictEqual(composerActions.rowByIndex(4).value(), null);
+    assert.deepEqual(composerActions.actionIds().sort(), [
+      "reply_as_new_topic",
+      "reply_to_topic",
+    ]);
   });
 
   test("replying to post - reply_to_topic", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
 
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
@@ -75,49 +95,46 @@ acceptance(`Composer Actions`, function (needs) {
     await composerActions.selectRowByValue("reply_to_topic");
 
     assert
-      .dom(".action-title .topic-link")
-      .hasText("Internationalization / localization");
-    assert
-      .dom(".action-title .topic-link")
-      .hasAttribute("href", "/t/internationalization-localization/280");
+      .dom(".composer-actions-trigger")
+      .includesText(i18n("composer.composer_actions.reply_to_topic.trigger"));
     assert
       .dom(".d-editor-input")
       .hasValue("test replying to topic when initially replied to post");
   });
 
-  test("replying to post - toggle_whisper for whisperers", async function (assert) {
+  test("toggle whisper via actions dropdown for whisperers", async function (assert) {
     updateCurrentUser({ admin: false, moderator: false });
-    const composerActions = selectKit(".composer-actions");
 
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
-    await fillIn(
-      ".d-editor-input",
-      "test replying as whisper to topic when initially not a whisper"
-    );
+
+    await click(".composer-actions-trigger");
 
     assert
-      .dom(".composer-actions svg.d-icon-far-eye-slash")
-      .doesNotExist("whisper icon is not visible");
-    assert
-      .dom(".composer-actions svg.d-icon-reply")
-      .exists("reply icon is visible");
+      .dom(".composer-toggle-whisper")
+      .exists("whisper toggle item is visible in dropdown");
 
-    await composerActions.expand();
-    await composerActions.selectRowByValue("toggle_whisper");
+    await click(".composer-toggle-whisper .d-toggle-switch__checkbox");
 
     assert
-      .dom(".composer-actions svg.d-icon-far-eye-slash")
-      .exists("whisper icon is visible");
+      .dom(
+        ".composer-toggle-whisper .d-toggle-switch__checkbox[aria-checked='true']"
+      )
+      .exists("whisper toggle is on after click");
+
+    await click(".composer-toggle-whisper .d-toggle-switch__checkbox");
+
     assert
-      .dom(".composer-actions svg.d-icon-reply")
-      .doesNotExist("reply icon is not visible");
+      .dom(
+        ".composer-toggle-whisper .d-toggle-switch__checkbox[aria-checked='false']"
+      )
+      .exists("whisper toggle is off after second click");
   });
 
   test("replying to post - reply_as_new_topic", async function (assert) {
     sinon.stub(Draft, "get").resolves({ draft: "", draft_sequence: 0 });
 
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
     const categoryChooser = selectKit(".title-wrapper .category-chooser");
     const categoryChooserReplyArea = selectKit(".reply-area .category-chooser");
     const quote = "test replying as new topic when initially replied to post";
@@ -136,34 +153,22 @@ acceptance(`Composer Actions`, function (needs) {
     await composerActions.selectRowByValue("reply_as_new_topic");
 
     assert.strictEqual(categoryChooserReplyArea.header().name(), "faq");
-    assert.dom(".action-title").hasText(i18n("topic.create_long"));
     assert.dom(".d-editor-input").includesValue(quote);
   });
 
   test("reply_as_new_topic without a new_topic draft", async function (assert) {
     await visit("/t/internationalization-localization/280");
     await click(".create.reply");
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
     await composerActions.expand();
     await composerActions.selectRowByValue("reply_as_new_topic");
     assert.dom(".dialog-body").doesNotExist();
   });
 
-  test("reply_as_new_topic without a permission to create topic", async function (assert) {
-    updateCurrentUser({ can_create_topic: false });
-    await visit("/t/internationalization-localization/280");
-    await click(".create.reply");
-    const composerActions = selectKit(".composer-actions");
-    await composerActions.expand();
-    assert
-      .dom(".composer-actions svg.d-icon-plus")
-      .doesNotExist("reply as new topic icon is not visible");
-  });
-
   test("reply_as_new_group_message", async function (assert) {
     await visit("/t/lorem-ipsum-dolor-sit-amet/130");
     await click(".create.reply");
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
     await composerActions.expand();
     await composerActions.selectRowByValue("reply_as_new_group_message");
 
@@ -171,8 +176,19 @@ acceptance(`Composer Actions`, function (needs) {
     assert.deepEqual(privateMessageUsers.header().value(), "foo,foo_group");
   });
 
+  test("reply_as_new_topic without a permission to create topic", async function (assert) {
+    updateCurrentUser({ can_create_topic: false });
+    await visit("/t/internationalization-localization/280");
+    await click(".create.reply");
+    const composerActions = composerActionsDropdown();
+    await composerActions.expand();
+    assert
+      .dom(".composer-actions-dropdown [data-action-id='reply_as_new_topic']")
+      .doesNotExist("reply as new topic option is not visible");
+  });
+
   test("interactions", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
     const quote = "Life is like riding a bicycle.";
 
     await visit("/t/short-topic-with-two-posts/54077");
@@ -181,177 +197,109 @@ acceptance(`Composer Actions`, function (needs) {
     await composerActions.expand();
     await composerActions.selectRowByValue("reply_to_topic");
 
-    assert.dom(".action-title").hasText("Short topic with two posts");
+    assert
+      .dom(".composer-actions-trigger")
+      .includesText(i18n("composer.composer_actions.reply_to_topic.trigger"));
     assert.dom(".d-editor-input").hasValue(quote);
 
     await composerActions.expand();
 
-    assert.strictEqual(
-      composerActions.rowByIndex(0).value(),
-      "reply_as_new_topic"
-    );
-    assert.strictEqual(composerActions.rowByIndex(1).value(), "reply_to_post");
-    assert.strictEqual(composerActions.rowByIndex(2).value(), "toggle_whisper");
-    assert.strictEqual(
-      composerActions.rowByIndex(3).value(),
-      "toggle_topic_bump"
-    );
-    assert.strictEqual(composerActions.rows().length, 4);
+    assert.deepEqual(composerActions.actionIds().sort(), [
+      "reply_as_new_topic",
+      "reply_to_post",
+    ]);
 
     await composerActions.selectRowByValue("reply_to_post");
     await composerActions.expand();
 
-    assert.dom(".action-title img.avatar").exists();
-    assert.dom(".action-title .user-link").hasText("tms");
+    assert.dom(".composer-actions-trigger").includesText("tms");
     assert.dom(".d-editor-input").hasValue(quote);
-    assert.strictEqual(
-      composerActions.rowByIndex(0).value(),
-      "reply_as_new_topic"
-    );
-    assert.strictEqual(composerActions.rowByIndex(1).value(), "reply_to_topic");
-    assert.strictEqual(composerActions.rowByIndex(2).value(), "toggle_whisper");
-    assert.strictEqual(
-      composerActions.rowByIndex(3).value(),
-      "toggle_topic_bump"
-    );
-    assert.strictEqual(composerActions.rows().length, 4);
+    assert.deepEqual(composerActions.actionIds().sort(), [
+      "reply_as_new_topic",
+      "reply_to_topic",
+    ]);
 
     await composerActions.selectRowByValue("reply_as_new_topic");
     await composerActions.expand();
 
-    assert.dom(".action-title").hasText(i18n("topic.create_long"));
     assert.dom(".d-editor-input").includesValue(quote);
-    assert.strictEqual(composerActions.rowByIndex(0).value(), "reply_to_post");
-    assert.strictEqual(composerActions.rowByIndex(1).value(), "reply_to_topic");
-    assert.strictEqual(
-      composerActions.rowByIndex(2).value(),
-      "toggle_unlisted"
-    );
-    assert.strictEqual(composerActions.rowByIndex(3).value(), "shared_draft");
-    assert.strictEqual(
-      composerActions.rowByIndex(4).value(),
-      "create_private_message"
-    );
-    assert.strictEqual(composerActions.rows().length, 5);
-  });
-
-  test("new topic in shared drafts category opens shared draft composer", async function (assert) {
-    await visit("/c/shared-drafts/24");
-
-    assert.dom("#create-topic").hasText(i18n("topic.create_shared_draft"));
-
-    await click("#create-topic");
-
-    assert
-      .dom("#reply-control .btn-primary.create .d-button-label")
-      .hasText(i18n("composer.create_shared_draft"));
-    assert
-      .dom(".composer-actions svg.d-icon-far-clipboard")
-      .exists("shared draft icon is visible");
-    assert.strictEqual(
-      selectKit(".category-chooser").header().value(),
-      null,
-      "shared drafts category is not selected as the destination category"
-    );
+    assert.deepEqual(composerActions.actionIds().sort(), [
+      "create_private_message",
+      "reply_to_post",
+      "reply_to_topic",
+      "shared_draft",
+    ]);
   });
 
   test("interactions - private message", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
 
     await visit("/t/internationalization-localization/280");
     await click('#post_4 a[data-user-card="shade"]');
     await click(".usercard-controls .compose-pm .btn-primary");
     await composerActions.expand();
 
-    assert.dom(".action-title").hasText(i18n("topic.private_message"));
-    assert.strictEqual(composerActions.rowByIndex(0).value(), "create_topic");
-    assert.strictEqual(composerActions.rows().length, 1);
+    assert.deepEqual(composerActions.actionIds().sort(), ["create_topic"]);
   });
 
-  test("replying to post - toggle_topic_bump", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
+  test("reply target link uses plain topic title when fancy title includes HTML", async function (assert) {
+    await visit("/t/short-topic-with-two-posts/54077");
+    await click(".create.reply");
 
+    await visit("/");
+    await waitFor(".composer-actions-reply-target-link__label");
+
+    assert
+      .dom(".composer-actions-reply-target-link__label")
+      .hasText(
+        "Short topic with two posts",
+        "renders the plain topic title in the reply target link"
+      );
+  });
+
+  test("toggle no-bump via actions dropdown", async function (assert) {
     await visit("/t/short-topic-with-two-posts/54077");
     await click("article#post_2 button.reply");
 
-    assert
-      .dom(".composer-actions svg.d-icon-anchor")
-      .doesNotExist("no-bump icon is not visible");
-    assert
-      .dom(".composer-actions svg.d-icon-reply")
-      .exists("reply icon is visible");
-
-    await composerActions.expand();
-    await composerActions.selectRowByValue("toggle_topic_bump");
+    await click(".composer-actions-trigger");
 
     assert
-      .dom(".composer-actions svg.d-icon-anchor")
-      .exists("no-bump icon is visible");
-    assert
-      .dom(".composer-actions svg.d-icon-reply")
-      .doesNotExist("reply icon is not visible");
+      .dom(".composer-toggle-no-bump")
+      .exists("no-bump toggle item is visible in dropdown");
 
-    await composerActions.expand();
-    await composerActions.selectRowByValue("toggle_topic_bump");
+    await click(".composer-toggle-no-bump .d-toggle-switch__checkbox");
 
     assert
-      .dom(".composer-actions svg.d-icon-anchor")
-      .doesNotExist("no-bump icon is not visible");
+      .dom(
+        ".composer-toggle-no-bump .d-toggle-switch__checkbox[aria-checked='true']"
+      )
+      .exists("no-bump toggle is on after click");
+
+    await click(".composer-toggle-no-bump .d-toggle-switch__checkbox");
+
     assert
-      .dom(".composer-actions svg.d-icon-reply")
-      .exists("reply icon is visible");
+      .dom(
+        ".composer-toggle-no-bump .d-toggle-switch__checkbox[aria-checked='false']"
+      )
+      .exists("no-bump toggle is off after second click");
   });
 
-  test("replying to post - whisper and no bump", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
-
-    await visit("/t/short-topic-with-two-posts/54077");
-    await click("article#post_2 button.reply");
-
-    assert
-      .dom(".composer-actions svg.d-icon-far-eye-slash")
-      .doesNotExist("whisper icon is not visible");
-    assert
-      .dom(".reply-details .whisper .d-icon-anchor")
-      .doesNotExist("no-bump icon is not visible");
-    assert
-      .dom(".composer-actions svg.d-icon-reply")
-      .exists("reply icon is visible");
-
-    await composerActions.expand();
-    await composerActions.selectRowByValue("toggle_topic_bump");
-    await composerActions.expand();
-    await composerActions.selectRowByValue("toggle_whisper");
-
-    assert
-      .dom(".composer-actions svg.d-icon-far-eye-slash")
-      .exists("whisper icon is visible");
-    assert
-      .dom(".reply-details .no-bump .d-icon-anchor")
-      .exists("no-bump icon is visible");
-    assert
-      .dom(".composer-actions svg.d-icon-reply")
-      .doesNotExist("reply icon is not visible");
-  });
-
-  test("replying to post as staff", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
-
+  test("replying to post as staff shows whisper + no-bump toggles in dropdown", async function (assert) {
     updateCurrentUser({ admin: true });
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
-    await composerActions.expand();
 
-    assert.strictEqual(composerActions.rows().length, 4);
-    assert.strictEqual(
-      composerActions.rowByIndex(3).value(),
-      "toggle_topic_bump"
-    );
+    await click(".composer-actions-trigger");
+
+    assert
+      .dom(".composer-toggle-whisper")
+      .exists("whisper toggle is visible for staff");
+    assert
+      .dom(".composer-toggle-no-bump")
+      .exists("no-bump toggle is visible for staff");
   });
 
-  test("replying to post as TL3 user", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
-
+  test("replying to post as TL3 user shows no toggles in dropdown", async function (assert) {
     updateCurrentUser({
       moderator: false,
       admin: false,
@@ -361,21 +309,18 @@ acceptance(`Composer Actions`, function (needs) {
     });
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
-    await composerActions.expand();
 
-    assert.strictEqual(composerActions.rows().length, 2);
-    Array.from(composerActions.rows()).forEach((row) => {
-      assert.notStrictEqual(
-        row.value,
-        "toggle_topic_bump",
-        "toggle button is not visible"
-      );
-    });
+    await click(".composer-actions-trigger");
+
+    assert
+      .dom(".composer-toggle-whisper")
+      .doesNotExist("whisper toggle is not visible for TL3 non-whisperer");
+    assert
+      .dom(".composer-toggle-no-bump")
+      .doesNotExist("no-bump toggle is not visible for TL3 non-whisperer");
   });
 
-  test("replying to post as TL4 user", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
-
+  test("replying to post as TL4 user shows no-bump toggle in dropdown", async function (assert) {
     updateCurrentUser({
       moderator: false,
       admin: false,
@@ -385,25 +330,145 @@ acceptance(`Composer Actions`, function (needs) {
     });
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
-    await composerActions.expand();
 
-    assert.strictEqual(composerActions.rows().length, 3);
-    assert.strictEqual(
-      composerActions.rowByIndex(2).value(),
-      "toggle_topic_bump"
-    );
+    await click(".composer-actions-trigger");
+
+    assert
+      .dom(".composer-toggle-no-bump")
+      .exists("no-bump toggle is visible for TL4");
   });
 
   test("editing post", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
 
     await visit("/t/internationalization-localization/280");
     await click("article#post_1 button.show-more-actions");
     await click("article#post_1 button.edit");
     await composerActions.expand();
 
-    assert.strictEqual(composerActions.rows().length, 1);
-    assert.strictEqual(composerActions.rowByIndex(0).value(), "reply_to_post");
+    assert.deepEqual(
+      composerActions.actionIds(),
+      [],
+      "no switch actions are offered while editing a post"
+    );
+  });
+
+  test("trigger shows correct icon for reply mode", async function (assert) {
+    await visit("/t/internationalization-localization/280");
+    await click("article#post_3 button.reply");
+
+    assert
+      .dom(".composer-actions-trigger svg.d-icon-share")
+      .exists("shows share icon when replying");
+  });
+
+  test("trigger shows correct icon for create topic mode", async function (assert) {
+    await visit("/");
+    await click("#create-topic");
+
+    assert
+      .dom(".composer-actions-trigger svg.d-icon-far-pen-to-square")
+      .exists("shows pen-to-square icon when creating topic");
+  });
+
+  test("trigger shows correct icon for editing", async function (assert) {
+    await visit("/t/internationalization-localization/280");
+    await click("article#post_1 button.show-more-actions");
+    await click("article#post_1 button.edit");
+
+    assert
+      .dom(".composer-actions-trigger svg.d-icon-pencil")
+      .exists("shows pencil icon when editing");
+  });
+
+  test("trigger shows correct label for reply mode", async function (assert) {
+    await visit("/t/internationalization-localization/280");
+    await click("article#post_3 button.reply");
+
+    assert
+      .dom(".composer-actions-trigger")
+      .includesText("codinghorror", "shows reply to post label");
+  });
+
+  test("trigger shows the private message label when replying in a PM", async function (assert) {
+    await visit("/t/lorem-ipsum-dolor-sit-amet/130");
+    await click(".create.reply");
+
+    assert
+      .dom(".composer-actions-trigger")
+      .includesText(
+        i18n("composer.composer_actions.reply_to_message.trigger"),
+        "shows the private message reply label, not the topic label"
+      );
+  });
+
+  test("trigger shows correct label for create topic mode", async function (assert) {
+    await visit("/");
+    await click("#create-topic");
+
+    assert
+      .dom(".composer-actions-trigger")
+      .includesText(
+        i18n("composer.composer_actions.create_topic.label"),
+        "shows create topic label"
+      );
+  });
+
+  test("create topic mode shows correct actions and unlisted toggle", async function (assert) {
+    const composer = this.owner.lookup("service:composer");
+    const composerActions = composerActionsDropdown();
+    updateCurrentUser({ admin: true });
+
+    await visit("/");
+    await click("#create-topic");
+    await composerActions.expand();
+
+    assert
+      .dom(
+        ".composer-actions-dropdown [data-action-id='create_private_message']"
+      )
+      .exists("shows create private message action");
+
+    assert
+      .dom(".composer-toggle-unlisted")
+      .exists("unlisted toggle is visible for staff in create topic mode");
+
+    await click(".composer-toggle-unlisted .d-toggle-switch__checkbox");
+    composer.model.set("noBump", true);
+
+    assert
+      .dom(
+        ".composer-toggle-unlisted .d-toggle-switch__checkbox[aria-checked='true']"
+      )
+      .exists("unlisted toggle is on after click");
+
+    await composerActions.selectRowByValue("create_private_message");
+
+    assert.false(
+      composer.model.unlistTopic,
+      "unlisted state is cleared when switching to a private message"
+    );
+    assert.false(
+      composer.model.noBump,
+      "no-bump state is cleared when switching to a private message"
+    );
+
+    await composerActions.expand();
+    assert
+      .dom(".composer-toggle-unlisted")
+      .doesNotExist("unlisted toggle is not shown in private message mode");
+  });
+
+  test("create topic mode does not show reply_as_new_topic", async function (assert) {
+    const composerActions = composerActionsDropdown();
+
+    await visit("/");
+    await click("#create-topic");
+    await composerActions.expand();
+
+    assert
+      .dom(".composer-actions-dropdown [data-action-id='reply_as_new_topic']")
+      .doesNotExist("does not show reply_as_new_topic in create topic mode");
   });
 });
 
@@ -415,7 +480,7 @@ function stubDraftResponse() {
   });
 }
 
-acceptance(`Composer Actions With New Topic Draft`, function (needs) {
+acceptance("Composer Actions With New Topic Draft", function (needs) {
   needs.user({ whisperer: true });
 
   needs.site({
@@ -443,7 +508,7 @@ acceptance(`Composer Actions With New Topic Draft`, function (needs) {
     await tags.expand();
     await tags.selectRowByName("monkey");
 
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
     await composerActions.expand();
     await composerActions.selectRowByValue("shared_draft");
 
@@ -456,8 +521,18 @@ acceptance(`Composer Actions With New Topic Draft`, function (needs) {
       .dom("#reply-control .btn-primary.create .d-button-label")
       .hasText(i18n("composer.create_shared_draft"));
     assert
-      .dom(".composer-actions svg.d-icon-far-clipboard")
+      .dom(".composer-actions-trigger svg.d-icon-far-clipboard")
       .exists("shared draft icon is visible");
+
+    await composerActions.expand();
+    assert
+      .dom(".composer-actions-dropdown [data-action-id='create_topic']")
+      .exists("can switch back to create topic from shared draft mode");
+    assert
+      .dom(
+        ".composer-actions-dropdown [data-action-id='create_private_message']"
+      )
+      .exists("can switch to PM from shared draft mode");
   });
 
   test("reply_as_new_topic with new_topic draft", async function (assert) {
@@ -466,7 +541,7 @@ acceptance(`Composer Actions With New Topic Draft`, function (needs) {
 
     stubDraftResponse();
 
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
     await composerActions.expand();
     await composerActions.selectRowByValue("reply_as_new_topic");
 
@@ -479,7 +554,7 @@ acceptance(`Composer Actions With New Topic Draft`, function (needs) {
   });
 });
 
-acceptance(`Prioritize Username`, function (needs) {
+acceptance("Prioritize Username", function (needs) {
   needs.user();
   needs.settings({
     prioritize_username_in_ux: true,
@@ -490,7 +565,7 @@ acceptance(`Prioritize Username`, function (needs) {
     await visit("/t/short-topic-with-two-posts/54079");
     await click("article#post_2 button.reply");
 
-    assert.dom(".action-title .user-link").hasText("james_john");
+    assert.dom(".composer-actions-trigger").includesText("james_john");
   });
 
   test("Quotes use username", async function (assert) {
@@ -505,7 +580,7 @@ acceptance(`Prioritize Username`, function (needs) {
   });
 });
 
-acceptance(`Prioritize Full Name`, function (needs) {
+acceptance("Prioritize Full Name", function (needs) {
   needs.user();
   needs.settings({
     prioritize_username_in_ux: false,
@@ -514,11 +589,11 @@ acceptance(`Prioritize Full Name`, function (needs) {
 
   test("Reply to post use full name", async function (assert) {
     await visit("/t/short-topic-with-two-posts/54079");
-    await click("article#post_3 button.reply");
+    await click("article#post_2 button.reply");
 
     assert
-      .dom(".action-title .user-link")
-      .hasHtml("&lt;h1&gt;Tim Stone&lt;/h1&gt;");
+      .dom(".composer-actions-trigger")
+      .includesText("james, john, the third");
   });
 
   test("Quotes use full name", async function (assert) {
@@ -544,7 +619,62 @@ acceptance(`Prioritize Full Name`, function (needs) {
   });
 });
 
-acceptance(`Prioritizing Name fall back`, function (needs) {
+acceptance("Slow Mode", function (needs) {
+  needs.user();
+  needs.pretender((server, helper) => {
+    server.get("/t/130.json", () => {
+      const json = cloneJSON(topicFixtures["/t/130.json"]);
+      // The 130 fixture is a PM; slow mode only applies to regular topics.
+      json.archetype = "regular";
+      json.slow_mode_seconds = 600;
+      json.slow_mode_enabled_until = "2040-01-01T04:00:00.000Z";
+      return helper.response(json);
+    });
+  });
+
+  test("trigger shows the regular reply label with the slow-mode icon", async function (assert) {
+    await visit("/t/internationalization-localization/130");
+    await click("article#post_1 button.reply");
+
+    assert
+      .dom(".composer-actions-trigger svg.d-icon-hourglass-start")
+      .exists("uses the slow-mode hourglass icon");
+    assert
+      .dom(".composer-actions-trigger")
+      .includesText(
+        i18n("composer.composer_actions.reply_to_topic.trigger"),
+        "falls back to the standard reply-to-topic trigger label"
+      );
+  });
+});
+
+acceptance("Private Messages", function (needs) {
+  needs.user();
+  needs.pretender((server, helper) => {
+    server.get("/t/280.json", () => {
+      const json = cloneJSON(topicFixtures["/t/280/1.json"]);
+      json.archetype = "private_message";
+      return helper.response(json);
+    });
+  });
+
+  test("dropdown reply-to-thread item uses the private message label", async function (assert) {
+    const composerActions = composerActionsDropdown();
+
+    await visit("/t/internationalization-localization/280");
+    await click("article#post_3 button.reply");
+    await composerActions.expand();
+
+    assert
+      .dom(".composer-actions-dropdown [data-action-id='reply_to_topic']")
+      .includesText(
+        i18n("composer.composer_actions.reply_to_message.label"),
+        "offers replying to the private message, not the topic"
+      );
+  });
+});
+
+acceptance("Prioritizing Name fall back", function (needs) {
   needs.user();
   needs.settings({
     prioritize_username_in_ux: false,
@@ -564,7 +694,7 @@ acceptance(`Prioritizing Name fall back`, function (needs) {
   });
 });
 
-acceptance(`Composer Actions - plugin registration`, function (needs) {
+acceptance("Composer Actions - plugin registration", function (needs) {
   needs.user();
 
   let lastInvocation;
@@ -590,32 +720,32 @@ acceptance(`Composer Actions - plugin registration`, function (needs) {
   });
 
   test("registered action appears in dropdown when condition is true", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
 
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
     await composerActions.expand();
 
     assert
-      .dom(`.composer-actions li[data-value="my_plugin_action"]`)
+      .dom(".composer-actions-dropdown [data-action-id='my_plugin_action']")
       .exists("the registered action is rendered");
   });
 
   test("registered action is hidden when condition is false", async function (assert) {
     showCustomAction = false;
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
 
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
     await composerActions.expand();
 
     assert
-      .dom(`.composer-actions li[data-value="my_plugin_action"]`)
+      .dom(".composer-actions-dropdown [data-action-id='my_plugin_action']")
       .doesNotExist("the registered action is not rendered");
   });
 
   test("selecting the registered action invokes its callback", async function (assert) {
-    const composerActions = selectKit(".composer-actions");
+    const composerActions = composerActionsDropdown();
 
     await visit("/t/internationalization-localization/280");
     await click("article#post_3 button.reply");
@@ -642,18 +772,18 @@ acceptance(`Composer Actions - plugin registration`, function (needs) {
 
     const errorStub = sinon.stub(console, "error");
     try {
-      const composerActions = selectKit(".composer-actions");
+      const composerActions = composerActionsDropdown();
 
       await visit("/t/internationalization-localization/280");
       await click("article#post_3 button.reply");
       await composerActions.expand();
-      await composerActions.collapse();
+      await click(".composer-actions-trigger");
       await composerActions.expand();
-      await composerActions.collapse();
+      await click(".composer-actions-trigger");
       await composerActions.expand();
 
       assert
-        .dom(`.composer-actions .select-kit-row[data-value="broken_action"]`)
+        .dom(".composer-actions-dropdown [data-action-id='broken_action']")
         .doesNotExist("broken action is hidden");
       const calls = errorStub
         .getCalls()
