+++ plugins/discourse-events/spec/system/composer/create_event_action_spec.rb
@@ -59,7 +59,7 @@ describe "Composer - Create event action" do
       I18n.t("js.discourse_post_event.composer.create_event_button"),
     )
 
-    composer.open_composer_actions_new
+    composer.open_composer_actions
     composer.select_action_by_id("create_regular_topic")
 
     expect(composer.button_label).to have_text(I18n.t("js.composer.create_topic"))
