+++ plugins/discourse-post-voting/test/javascripts/acceptance/composer-test.js
@@ -42,15 +42,12 @@ acceptance("composer", function (needs) {
     await categoryChooser.expand();
     await categoryChooser.selectRowByValue(2);
 
-    const composerActions = selectKit(".composer-actions");
-    await composerActions.expand();
-    await composerActions.selectKitSelectRowByName(
-      i18n("composer.composer_actions.create_as_post_voting.label")
-    );
+    await click(".composer-actions-trigger");
+    await click("[data-action-id='togglePostVoting']");
 
     assert
-      .dom(".action-title")
-      .hasText(
+      .dom(".composer-actions-trigger")
+      .includesText(
         i18n("composer.create_post_voting.label"),
         "displays the right composer action title when creating Post Voting topic"
       );
@@ -62,22 +59,18 @@ acceptance("composer", function (needs) {
         "displays the right label for composer create button"
       );
 
-    await composerActions.expand();
-    await composerActions.selectKitSelectRowByName(
-      i18n("composer.composer_actions.remove_as_post_voting.label")
-    );
+    await click(".composer-actions-trigger");
+    await click("[data-action-id='togglePostVoting']");
 
     assert
-      .dom(".action-title")
+      .dom(".composer-actions-trigger")
       .doesNotIncludeText(
         i18n("composer.create_post_voting.label"),
         "reverts to original composer title when post voting format is disabled"
       );
 
-    await composerActions.expand();
-    await composerActions.selectKitSelectRowByName(
-      i18n("composer.composer_actions.create_as_post_voting.label")
-    );
+    await click(".composer-actions-trigger");
+    await click("[data-action-id='togglePostVoting']");
 
     await fillIn("#reply-title", "this is some random topic title");
     await fillIn(".d-editor-input", "this is some random body");
@@ -95,15 +88,13 @@ acceptance("composer", function (needs) {
     await visit("/");
     await click("#create-topic");
 
-    assert.dom(".action-title").hasText(i18n("topic.create_long"));
-
     const categoryChooser = selectKit(".category-chooser");
     await categoryChooser.expand();
     await categoryChooser.selectRowByValue(2);
 
     assert
-      .dom(".action-title")
-      .hasText(i18n("composer.create_post_voting.label"));
+      .dom(".composer-actions-trigger")
+      .includesText(i18n("composer.create_post_voting.label"));
   });
 
   test("Creating new topic in category with only_post_voting_in_this_category enabled", async function (assert) {
@@ -113,16 +104,13 @@ acceptance("composer", function (needs) {
     await visit("/");
     await click("#create-topic");
 
-    assert.dom(".action-title").hasText(i18n("topic.create_long"));
-
     const categoryChooser = selectKit(".category-chooser");
     await categoryChooser.expand();
 
     await categoryChooser.selectRowByValue(2);
-    const newTopicType = selectKit(".dropdown-select-box");
-    await newTopicType.expand();
+
     assert
-      .dom(".action-title")
-      .hasText(i18n("composer.create_post_voting.label"));
+      .dom(".composer-actions-trigger")
+      .includesText(i18n("composer.create_post_voting.label"));
   });
 });
