+++ plugins/discourse-presence/test/javascripts/acceptance/discourse-presence-test.js
@@ -107,9 +107,8 @@ acceptance("Discourse Presence Plugin", function (needs) {
       "publishes reply presence when typing"
     );
 
-    const menu = selectKit(".composer-actions");
-    await menu.expand();
-    await menu.selectRowByValue("toggle_whisper");
+    await click(".composer-actions-trigger");
+    await click(".composer-toggle-whisper .d-toggle-switch__checkbox");
 
     assert.deepEqual(
       presentUserIds("/discourse-presence/reply/280"),
