+++ spec/system/private_message_spec.rb
@@ -29,7 +29,7 @@ describe "Private Message" do
       visit "/new-topic"
       expect(composer).to be_opened
 
-      composer.open_composer_actions_new
+      composer.open_composer_actions
       composer.select_action_by_id("create_private_message")
 
       expect(composer.button_label).to have_text(I18n.t("js.composer.create_pm"))
@@ -41,7 +41,7 @@ describe "Private Message" do
       pm_post_obj.reply
       expect(composer).to be_opened
 
-      composer.open_composer_actions_new
+      composer.open_composer_actions
       expect(composer).to have_no_action_id("create_private_message")
     end
 
@@ -50,12 +50,12 @@ describe "Private Message" do
       expect(composer).to be_opened
 
       # Switch to personal message
-      composer.open_composer_actions_new
+      composer.open_composer_actions
       composer.select_action_by_id("create_private_message")
       expect(composer.button_label).to have_text(I18n.t("js.composer.create_pm"))
 
       # Switch back to topic
-      composer.open_composer_actions_new
+      composer.open_composer_actions
       composer.select_action_by_id("create_topic")
       expect(composer.button_label).to have_text(I18n.t("js.composer.create_topic"))
     end
